backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
email-validator==2.1.0

# Database (MySQL)
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for TMDB API)
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning (for recommendations)
scikit-learn==1.3.2
scipy==1.11.4

# Advanced ML Features
gensim==4.3.2  # Word2Vec for semantic similarity
textblob==0.17.1  # Sentiment analysis
nltk==3.8.1  # Natural language processing

# Redis (Optional - for caching)
redis==5.0.1
hiredis==2.2.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# System Monitoring
psutil==5.9.6

# Additional Testing
pytest-cov==4.1.0
pytest-mock==3.12.0
